backend/services/ai-agent-service/requirements-simple.txt
google-generativeai==0.8.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
python-dotenv==1.0.0